03-sqlalchemy-alembic/pyproject.toml
[tool.poetry]
name = "03-sqlalchemy-alembic"
version = "0.1.0"
description = ""
authors = ["Rashmin Dungrani <rashmindungrani@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
orjson = "^3.10.3"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
